Fixes for buliding for 32bit targets:   * u32/u64/uword mismatches   * pointer-to-int fixes   * printf formatting issues   * issues with incorrect "ULL" and related suffixes   * structure alignment and padding issues

+++ src/vcl/ldp.c
@@ -554,9 +554,10 @@ writev (int fd, const struct iovec * iov, int iovcnt)
 	    {
 	      if (LDP_DEBUG > 4)
 		printf ("%s:%d: LDP<%d>: fd %d (0x%x): calling %s() [%d]: "
-			"sid %u (0x%x), buf %p, nbytes %ld, total %ld",
+			"sid %u (0x%x), buf %p, nbytes %lu, total %ld",
 			__func__, __LINE__, getpid (), fd, fd, func_str,
-			i, sid, sid, iov[i].iov_base, iov[i].iov_len, total);
+			i, sid, sid, iov[i].iov_base,
+			(unsigned long) iov[i].iov_len, (long) total);
 
 	      rv = vppcom_session_write (sid, iov[i].iov_base,
 					 iov[i].iov_len);
@@ -569,9 +570,9 @@ writev (int fd, const struct iovec * iov, int iovcnt)
 		    {
 		      if (LDP_DEBUG > 4)
 			printf ("%s:%d: LDP<%d>: fd %d (0x%x): "
-				"rv (%d) < iov[%d].iov_len (%ld)",
+				"rv (%d) < iov[%d].iov_len (%lu)",
 				__func__, __LINE__, getpid (), fd, fd,
-				rv, i, iov[i].iov_len);
+				rv, i, (unsigned long) iov[i].iov_len);
 		      break;
 		    }
 		}
@@ -608,12 +609,12 @@ writev (int fd, const struct iovec * iov, int iovcnt)
 	  fprintf (stderr,
 		   "%s:%d: LDP<%d>: ERROR: fd %d (0x%x): %s() failed! "
 		   "rv %ld, errno = %d\n", __func__, __LINE__, getpid (), fd,
-		   fd, func_str, size, errno_val);
+		   fd, func_str, (long) size, errno_val);
 	  errno = errno_val;
 	}
       else
 	printf ("%s:%d: LDP<%d>: fd %d (0x%x): returning %ld\n",
-		__func__, __LINE__, getpid (), fd, fd, size);
+		__func__, __LINE__, getpid (), fd, fd, (long) size);
     }
   return size;
 }
@@ -1374,8 +1375,8 @@ ldp_copy_ep_to_sockaddr (__SOCKADDR_ARG addr, socklen_t * __restrict len,
 	  sa_len = sizeof (struct sockaddr_in6) - sizeof (struct in6_addr);
 	  copy_len = *len - sa_len;
 	  if (copy_len > 0)
-	    memcpy (((struct sockaddr_in6 *) addr)->sin6_addr.
-		    __in6_u.__u6_addr8, ep->ip, copy_len);
+	    memcpy (((struct sockaddr_in6 *) addr)->sin6_addr.__in6_u.
+		    __u6_addr8, ep->ip, copy_len);
 	  break;
 
 	default:
